Fixed unit test compilation on Windows

+++ tests/unit/core/hpe/Process.cpp
@@ -72,7 +72,7 @@ TEST_F(ProcessTest, Mode) {
 	size_t count;
 	Process *proc = createProcess();
 	count = proc->nAccelerators();
-	ASSERT_EQ(gmacSuccess, proc->migrate(count-1));
+	ASSERT_EQ(gmacSuccess, proc->migrate(static_cast<int>(count) - 1));
 	Mode &mode = Thread::getCurrentMode();
 	Accelerator &acc = mode.getAccelerator();
 	ASSERT_TRUE(acc.id() == count-1);
